Fix the broken test different_extensions/.babelrc
Result is arguable. It looks like now prettier will automatically convert to
single quotes and not convert the .babelrc to a proper JSON-format, and instead
use quote-less keys and single-quoted values.

It does not look aligned with the other plugins to me, but it *is* prettier's
behavior, so this seems like the new correct assumption for now.